Split object command execute checks
