+++ doc/reference/file-map.md
@@ -50,6 +50,7 @@
 - `include/app/workspace_layout_types.h`
 - `include/app/workspace_service_types.h`
 - `src/commands/command_executor.c`
+- `src/commands/command_object_checks.c`
 - `src/commands/command_object_ops.c`
 - `src/commands/command_layer_ops.c`
 - `src/commands/command_transaction.c`
